smart-dataset-generator-backend/requirements.txt
# Smart Dataset Generator Backend Requirements

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# File handling
openpyxl==3.1.2

# CORS support
python-multipart==0.0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For enhanced data processing
# scikit-learn==1.3.2
# matplotlib==3.8.2
# seaborn==0.13.0
